improvement : add floating selected menu & improve loading product detail
