+++ package.json
@@ -39,6 +39,7 @@
     "react-hook-form": "^7.51.1",
     "react-lazily": "^0.9.2",
     "react-router-dom": "6.11.2",
+    "recoil": "^0.7.7",
     "superjson": "^1.13.0",
     "ts-pattern": "^5.0.8",
     "zod": "^3.22.4"
